Generate predictions from a trained classification model

This function loads a trained classification model from an .rds file and uses it to make predictions on a test dataset. It can also return class probabilities but it is disabled by default.

Usage

testing(model_file, test_data, return_probs = FALSE)

Arguments

model_file

Character. Path to the saved .rds model file

test_data

Data frame or tibble. Test dataset with both regular and target feature

return_probs

Logical. If TRUE, the function also returns class probabilities

Value

If return_probs = FALSE, returns a character vector of predicted class labels. If return_probs = TRUE, returns a named list with:

predictions: A character vector of predicted class labels

probabilities: A matrix or data frame of class probabilities

Examples

if (FALSE) { # \dontrun{
testing("model.rds", test_data)
testing("model.rds", test_data, return_probs = TRUE)
} # }